+++ src/main/java/frc/robot/RobotContainer.java
@@ -8,20 +8,9 @@ import edu.wpi.first.wpilibj.Joystick;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.button.*;
 import frc.robot.commands.*;
-import frc.robot.commands.ClampCmd;
 import frc.robot.commands.Claw.*;
 import frc.robot.commands.OneBar.*;
-import frc.robot.commands.ReleaseCmd;
-import frc.robot.commands.ShiftLeftCmd;
-import frc.robot.commands.ShiftRightCmd;
-import frc.robot.commands.StopClawCmd;
 import frc.robot.commands.Wrist.*;
-import frc.robot.commands.Wrist.LowerCmd;
-import frc.robot.commands.Wrist.RaiseCmd;
-import frc.robot.commands.Wrist.RotLeftCmd;
-import frc.robot.commands.Wrist.RotRightCmd;
-import frc.robot.commands.Wrist.StopRaiseCmd;
-import frc.robot.commands.Wrist.StopRotCmd;
 import frc.robot.subsystems.*;
 import frc.robot.subsystems.ClawSubsystem;
 import frc.robot.subsystems.WristSubsystem;
