+++ src/main/java/frc/robot/commands/Wrist/RotLeft90Cmd.java
@@ -2,30 +2,35 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.Wrist;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.subsystems.ClawSubsystem;
+import frc.robot.subsystems.WristSubsystem;
 
-public class ShiftLeftCmd extends CommandBase {
-  /** Creates a new ShiftLeftCmd. */
-  private ClawSubsystem claw;
+public class RotLeft90Cmd extends CommandBase {
+  /** Creates a new RotLeft90Cmd. */
+  private WristSubsystem wrist;
 
-  public ShiftLeftCmd(ClawSubsystem claws) {
-    claw = claws;
-    addRequirements(claws);
+  public RotLeft90Cmd(WristSubsystem wrists) {
+
+    wrist = wrists;
+    addRequirements(wrists);
   }
 
   // Called when the command is initially scheduled.
   @Override
   public void initialize() {
 
-    claw.shiftleft();
+    wrist.resetrot();
   }
 
   // Called every time the scheduler runs while the command is scheduled.
   @Override
-  public void execute() {}
+  public void execute() {
+
+    // wrist.encWrisrotleft90();
+    wrist.rotWrist90Left();
+  }
 
   // Called once the command ends or is interrupted.
   @Override
@@ -34,6 +39,14 @@ public class ShiftLeftCmd extends CommandBase {
   // Returns true when the command should end.
   @Override
   public boolean isFinished() {
-    return false;
+
+    double rotation = wrist.getrotations() * 360;
+    System.out.println("Rotations " + rotation);
+    if (rotation < -90) {
+      wrist.stoprot();
+      return true;
+    } else {
+      return false;
+    }
   }
 }
