+++ src/main/java/frc/robot/commands/Wrist/RotRight90Cmd.java
@@ -2,25 +2,22 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands.Claw;
+package frc.robot.commands.Wrist;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.subsystems.WristSubsystem;
 
-public class RotLeft90Cmd extends CommandBase {
-  /** Creates a new RotLeft90Cmd. */
+public class RotRight90Cmd extends CommandBase {
+  /** Creates a new RotRight90Cmd. */
   private WristSubsystem wrist;
 
-  public RotLeft90Cmd(WristSubsystem wrists) {
-
+  public RotRight90Cmd(WristSubsystem wrists) {
     wrist = wrists;
     addRequirements(wrists);
   }
-
   // Called when the command is initially scheduled.
   @Override
   public void initialize() {
-
     wrist.resetrot();
   }
 
@@ -28,8 +25,7 @@ public class RotLeft90Cmd extends CommandBase {
   @Override
   public void execute() {
 
-    // wrist.encWrisrotleft90();
-    wrist.rotWrist90Left();
+    wrist.rotWristRight90();
   }
 
   // Called once the command ends or is interrupted.
@@ -39,10 +35,9 @@ public class RotLeft90Cmd extends CommandBase {
   // Returns true when the command should end.
   @Override
   public boolean isFinished() {
-
     double rotation = wrist.getrotations() * 360;
     System.out.println("Rotations " + rotation);
-    if (rotation < -90) {
+    if (rotation > 90) {
       wrist.stoprot();
       return true;
     } else {
